% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineDupCols.R
\name{combineDupCols}
\alias{combineDupCols}
\title{Combine duplicate columns}
\usage{
combineDupCols(x, cols, new_name)
}
\arguments{
\item{x}{A dataframe}

\item{cols}{A vector of names of duplicate columns, as strings}

\item{new_name}{The new name of the combined column, as a string}
}
\description{
Combine duplicate columns
}
