% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configSkeleton.R
\name{configSkeleton}
\alias{configSkeleton}
\title{Get encoded values by trait ID}
\usage{
configSkeleton(db, dbgap_trait_accessions, file = "RENAME_ME_config.xml")
}
\arguments{
\item{db}{A database connection}

\item{dbgap_trait_accessions}{A vector of dbGaP variable accessions, as integers}

\item{file}{File path and filename for exported XML config file}
}
\description{
Get encoded values by trait ID
}
