% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcPlot.R
\name{qcPlot}
\alias{qcPlot}
\title{Fit a model and plot the residuals}
\usage{
qcPlot(dat, outcome, covarsF, covarsN, byVars, type, subj.include = NULL,
  title = NULL, skewed)
}
\arguments{
\item{dat}{A dataframe}

\item{outcome}{name in dat for the quantitative trait}

\item{covarsF}{names in dat for covariates that are to be factors}

\item{covarsN}{names in dat for other covariates}

\item{byVars}{grouping variables of interest}

\item{type}{type of plot: should be "box" or "violin"}

\item{subj.include}{subject IDs for subjects to include; if NULL, all subjects in dat will be used}

\item{title}{brief text to prepend to automatic title}
}
\description{
Fit a model and plot the residuals
}
