% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDats.R
\name{readDats}
\alias{readDats}
\title{Loads datasets returned by harmonization. Requires .RData files to return dataset named `dat`}
\usage{
readDats(file_list)
}
\arguments{
\item{file_list}{A list of files to be passed to `load()`}
}
\description{
Loads datasets returned by harmonization. Requires .RData files to return dataset named `dat`
}
