% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderInputUnitNode.R
\name{renderInputUnitNode}
\alias{renderInputUnitNode}
\title{Builds an input unit node for a phenotype harmonization configuration file}
\usage{
renderInputUnitNode(unit_name, stids = NULL, atids = NULL, htsids = NULL,
  btids = NULL, hfs)
}
\arguments{
\item{unit_name}{Name of the harmonization unit}

\item{stids}{A vector of source_trait_ids}

\item{atids}{A vector of source_trait_ids of phenotypes used to calculate age}

\item{htsids}{A vector of harmonized_trait_set_ids}

\item{btids}{A vector of source_trait_ids used for separating subjects/observations into batches}

\item{hf}{Path to the harmonization function for this unit}
}
\description{
Builds an input unit node for a phenotype harmonization configuration file
}
