% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderMetaNode.R
\name{renderMetaNode}
\alias{renderMetaNode}
\title{Creates the metadata node for a phenotype harmonization file}
\usage{
renderMetaNode(name, description, data_type, unit, encoded, qc_doc,
  is_demographic = FALSE, is_longitudinal = FALSE, has_batch = FALSE)
}
\arguments{
\item{name}{Name of the harmonized trait}

\item{description}{Description of the harmonized trait}

\item{data_type}{Data type of the output}

\item{unit}{Unit of the harmonized trait}

\item{encoded}{A named vector of encoded values}

\item{qc_doc}{File path to the accompanying QC document}

\item{is_demographic}{Flags output as a demographic trait}

\item{is_longitudinal}{Flags output as a longitudinal trait and allows multiple observations per topmed_subject_id}
}
\value{
The metadata node for a phenotype harmonization configuration file
}
\description{
Creates the metadata node for a phenotype harmonization file
}
