% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderOutputNode.R
\name{renderOutputNode}
\alias{renderOutputNode}
\title{Create the output node for a phenotype harmonization configuration file}
\usage{
renderOutputNode(output_dir, output_prefix)
}
\arguments{
\item{output_dir}{Output directory}

\item{output_prefix}{Prefix for output files}
}
\value{
Output node for a phenotype harmonization configuration file
}
\description{
Create the output node for a phenotype harmonization configuration file
}
