% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbcCheck.r
\name{wbcCheck}
\alias{wbcCheck}
\title{Check white blood cell measurements against the differential measurements}
\usage{
wbcCheck(data, wbc = "wbc_ncnc_bld", lymph = "lymphocyte_ncnc_bld",
  mono = "monocyte_ncnc_bld", neutro = "neutrophil_ncnc_bld",
  eosin = "eosinophil_ncnc_bld", baso = "basophil_ncnc_bld")
}
\arguments{
\item{data}{a dataframe}

\item{wbc}{The name of the column with WBC measurements}

\item{lymph}{The name of the column with lymphocyte measurements}

\item{mono}{The name of the column with monocyte measurements}

\item{neutro}{The name of the column with neutrophil measurements}

\item{eosin}{The name of the column with eosinophil measurements}

\item{baso}{The name of the column with basophil measurements}
}
\description{
Check white blood cell measurements against the differential measurements
}
