% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configSkeleton.R
\name{configSkeleton}
\alias{configSkeleton}
\title{Construct the skeleton for the XML configuration file for the harmonization batch. Takes named lists as arguments}
\usage{
configSkeleton(source_trait_ids, age_trait_ids, harmon_functions, name, qc_doc,
  output_dir = "", output_prefix = "output", description = "",
  data_type = "", unit)
}
\arguments{
\item{dbgap_trait_accessions}{A vector of dbGaP variable accessions, as integers}

\item{file}{File path and filename for exported XML config file}
}
\description{
Construct the skeleton for the XML configuration file for the harmonization batch. Takes named lists as arguments
}
